+++ packages/ad/ad.js
@@ -6,6 +6,12 @@ import { pbjs as pbjsConfig } from "./config";
 
 import Placeholder from "./placeholder";
 
+const styles = StyleSheet.create({
+  children: {
+    flex: 1
+  }
+});
+
 class Ad extends Component {
   static hasDifferentOrigin(url, baseUrl) {
     return url && url.indexOf(baseUrl) === -1 && url.indexOf("://") > -1;
@@ -69,11 +75,12 @@ class Ad extends Component {
         <head>
           <style>
             body {
-              margin: 0 auto;
               display: table;
               height: 100%;
-              width: 100%;
+              margin: 0 auto;
+              overflow:hidden;
               text-align: center;
+              width: 100%;
             }
             div#${this.props.code} {
               display: table-cell;
@@ -165,27 +172,39 @@ class Ad extends Component {
       </html>
       `;
 
-    return (
-      <View style={this.props.style}>
-        <Placeholder
+    const placeholderComponent = !this.state.adReady
+      ? <Placeholder
           width={this.config.maxSizes.width}
           height={this.config.maxSizes.height}
-          style={{
-            display: this.state.adReady ? "none" : "flex"
-          }}
+          style={styles.children}
         />
-        <WebView
-          ref={ref => {
-            this.webview = ref;
-          }}
-          source={{ html, baseUrl: this.props.baseUrl }}
-          style={{
-            height: this.config.maxSizes.height + this.viewBorder,
-            display: this.state.adReady ? "flex" : "none"
-          }}
-          baseUrl={this.props.baseUrl}
-          onNavigationStateChange={this.handleNavigationChange}
-        />
+      : null;
+
+    const webviewStyles = !this.state.adReady
+      ? {
+          width: 0,
+          height: 0
+        }
+      : {
+          height: this.config.maxSizes.height + this.viewBorder
+        };
+
+    const webviewComponent = (
+      <WebView
+        ref={ref => {
+          this.webview = ref;
+        }}
+        source={{ html, baseUrl: this.props.baseUrl }}
+        style={[webviewStyles, styles.children]}
+        baseUrl={this.props.baseUrl}
+        onNavigationStateChange={this.handleNavigationChange}
+      />
+    );
+
+    return (
+      <View style={[this.props.style]}>
+        {webviewComponent}
+        {placeholderComponent}
       </View>
     );
   }
@@ -197,7 +216,7 @@ Ad.propTypes = {
   code: PropTypes.string.isRequired,
   section: PropTypes.string.isRequired,
   baseUrl: PropTypes.string,
-  style: React.PropTypes.instanceOf(StyleSheet)
+  style: PropTypes.instanceOf(StyleSheet)
 };
 
 Ad.defaultProps = {