+++ packages/author-head/author-head.stories.js
@@ -5,7 +5,10 @@ import AuthorHead from "./author-head";
 
 const data = require("./fixtures/profile.json");
 
-const story = m => <View style={{ padding: 20 }}>{m}</View>;
+const story = m =>
+  <View style={{ padding: 20 }}>
+    {m}
+  </View>;
 
 storiesOf("AuthorHead", module).add("Full Header", () =>
   story(<AuthorHead {...data} />)