+++ packages/link/link.js
@@ -4,12 +4,17 @@ import { Text, Linking } from "react-native";
 import { defaultProps, propTypes } from "./link.proptypes";
 
 const styles = {
-  color: "#069"
+  color: "#069",
+  textDecorationLine: "underline"
 };
 
-export default function Link({ url, style, children }) {
+export default function Link(props) {
+  const { children, style, url } = props;
+
+  const { onPress = () => Linking.openURL(url) } = props;
+
   return (
-    <Text onPress={() => Linking.openURL(url)} style={[styles, style]}>
+    <Text {...props} onPress={onPress} style={[styles, style]}>
       {children}
     </Text>
   );