+++ packages/provider/provider.stories.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
+import { withPageState } from "@times-components/pagination";
+
 import {
   ApolloClient,
   ApolloProvider,
@@ -9,30 +11,20 @@ import {
   gql
 } from "react-apollo";
 
-import { withApolloProvider } from "storybook-addon-apollo-graphql";
-
 import connectGraphql, { AuthorProfileProvider } from "./provider.js";
 
-const Component = props => <Text>{JSON.stringify(props, null, 2)}</Text>;
+const Component = props =>
+  <Text>
+    {JSON.stringify(props, null, 2)}
+  </Text>;
 
-const schema = `
-    type Query {
-        random: Int!
+const query = gql`
+  {
+    author(slug: "fiona-hamilton") {
+      name
     }
+  }
 `;
-
-const root = {
-  random: () => Math.floor(Math.random() * 10)
-};
-
-const query = gql`{ random }`;
-
-storiesOf("Provider", module)
-  .addDecorator(withApolloProvider({ schema, root }))
-  .add("Provider", () => {
-    const WithData = connectGraphql(query)(Component);
-    return <WithData />;
-  });
 
 const fragmentMatcher = new IntrospectionFragmentMatcher({
   introspectionQueryResultData: {
@@ -64,13 +56,24 @@ const client = new ApolloClient({
   fragmentMatcher
 });
 
+storiesOf("Provider", module).add("Provider", () => {
+  const WithData = connectGraphql(query)(Component);
+  return (
+    <ApolloProvider client={client}>
+      <WithData />
+    </ApolloProvider>
+  );
+});
+
+const AuthorProfileWithPageState = withPageState(AuthorProfileProvider);
 storiesOf("Provider", module).add("AuthorProfileProvider", () =>
   <ApolloProvider client={client}>
-    <AuthorProfileProvider
+    <AuthorProfileWithPageState
+      generatePageLink={page => `https://www.thetimes.co.uk?page=${page}`}
+      imageRatio="3:2"
       slug="fiona-hamilton"
       page={1}
-      pageSize={10}
-      imageRatio="3:2"
+      pageSize={3}
     />
   </ApolloProvider>
 );