+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.12"></a>
+## [0.2.12](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.9...@times-components/article-summary@0.2.12) (2017-08-22)
+
+
+
+
+<a name="0.2.11"></a>
+## [0.2.11](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.9...@times-components/article-summary@0.2.11) (2017-08-22)
+
+
+
+
+<a name="0.2.10"></a>
+## [0.2.10](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.9...@times-components/article-summary@0.2.10) (2017-08-21)
+
+
+
+
+<a name="0.2.9"></a>
+## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.8...@times-components/article-summary@0.2.9) (2017-08-21)
+
+
+
+
+<a name="0.2.8"></a>
+## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.7...@times-components/article-summary@0.2.8) (2017-08-21)
+
+
+### Bug Fixes
+
+* date-publication add enum for publication values ([98fa88e](https://github.com/newsuk/times-components/commit/98fa88e))
+
+
+
+
+<a name="0.2.7"></a>
+## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.6...@times-components/article-summary@0.2.7) (2017-08-18)
+
+
+
+
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.5...@times-components/article-summary@0.2.6) (2017-08-18)
+
+
+
+
+<a name="0.2.5"></a>
+## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.4...@times-components/article-summary@0.2.5) (2017-08-18)
+
+
+
+
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.3...@times-components/article-summary@0.2.4) (2017-08-17)
+
+
+
+
 <a name="0.2.3"></a>
 ## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.2.2...@times-components/article-summary@0.2.3) (2017-08-17)