+++ packages/article/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.4.2"></a>
+## [0.4.2](https://github.com/newsuk/times-components/compare/@times-components/article@0.4.1...@times-components/article@0.4.2) (2017-08-22)
+
+
+
+
+<a name="0.4.1"></a>
+## [0.4.1](https://github.com/newsuk/times-components/compare/@times-components/article@0.4.0...@times-components/article@0.4.1) (2017-08-17)
+
+
+
+
 <a name="0.4.0"></a>
 # [0.4.0](https://github.com/newsuk/times-components/compare/@times-components/article@0.1.0...@times-components/article@0.4.0) (2017-08-11)