+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.5.1"></a>
+## [0.5.1](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.5.0...@times-components/author-profile@0.5.1) (2017-08-22)
+
+
+
+
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.8...@times-components/author-profile@0.5.0) (2017-08-22)
+
+
+### Features
+
+* author profile provider component ([57cbb79](https://github.com/newsuk/times-components/commit/57cbb79))
+
+
+
+
+<a name="0.4.8"></a>
+## [0.4.8](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.7...@times-components/author-profile@0.4.8) (2017-08-21)
+
+
+
+
+<a name="0.4.7"></a>
+## [0.4.7](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.6...@times-components/author-profile@0.4.7) (2017-08-21)
+
+
+
+
+<a name="0.4.6"></a>
+## [0.4.6](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.5...@times-components/author-profile@0.4.6) (2017-08-21)
+
+
+### Bug Fixes
+
+* date-publication add enum for publication values ([98fa88e](https://github.com/newsuk/times-components/commit/98fa88e))
+
+
+
+
+<a name="0.4.5"></a>
+## [0.4.5](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.4...@times-components/author-profile@0.4.5) (2017-08-18)
+
+
+
+
+<a name="0.4.4"></a>
+## [0.4.4](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.3...@times-components/author-profile@0.4.4) (2017-08-18)
+
+
+
+
+<a name="0.4.3"></a>
+## [0.4.3](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.2...@times-components/author-profile@0.4.3) (2017-08-18)
+
+
+
+
+<a name="0.4.2"></a>
+## [0.4.2](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.1...@times-components/author-profile@0.4.2) (2017-08-17)
+
+
+
+
 <a name="0.4.1"></a>
 ## [0.4.1](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.4.0...@times-components/author-profile@0.4.1) (2017-08-17)