+++ packages/date-publication/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.0.7"></a>
+## [0.0.7](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.0.6...@times-components/date-publication@0.0.7) (2017-08-21)
+
+
+### Bug Fixes
+
+* date-publication add enum for publication values ([98fa88e](https://github.com/newsuk/times-components/commit/98fa88e))
+
+
+
+
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.0.5...@times-components/date-publication@0.0.6) (2017-08-18)
+
+
+
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.0.4...@times-components/date-publication@0.0.5) (2017-08-17)
+
+
+
+
 <a name="0.0.4"></a>
 ## 0.0.4 (2017-08-11)