+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.4.1"></a>
+## [0.4.1](https://github.com/newsuk/times-components/compare/@times-components/markup@0.4.0...@times-components/markup@0.4.1) (2017-08-22)
+
+
+
+
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/newsuk/times-components/compare/@times-components/markup@0.3.3...@times-components/markup@0.4.0) (2017-08-22)
+
+
+### Features
+
+* author profile provider component ([57cbb79](https://github.com/newsuk/times-components/commit/57cbb79))
+
+
+
+
+<a name="0.3.3"></a>
+## [0.3.3](https://github.com/newsuk/times-components/compare/@times-components/markup@0.3.2...@times-components/markup@0.3.3) (2017-08-21)
+
+
+
+
+<a name="0.3.2"></a>
+## [0.3.2](https://github.com/newsuk/times-components/compare/@times-components/markup@0.3.1...@times-components/markup@0.3.2) (2017-08-21)
+
+
+### Bug Fixes
+
+* link - add underline to links ([f1ee6c6](https://github.com/newsuk/times-components/commit/f1ee6c6))
+
+
+
+
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/newsuk/times-components/compare/@times-components/markup@0.3.0...@times-components/markup@0.3.1) (2017-08-18)
+
+
+
+
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/newsuk/times-components/compare/@times-components/markup@0.2.1...@times-components/markup@0.3.0) (2017-08-18)
+
+
+### Features
+
+* add link component to markup builder ([#151](https://github.com/newsuk/times-components/issues/151)) ([6429a16](https://github.com/newsuk/times-components/commit/6429a16))
+
+
+
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/markup@0.2.0...@times-components/markup@0.2.1) (2017-08-17)
+
+
+
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/markup@0.1.2...@times-components/markup@0.2.0) (2017-08-17)